Draws a network representation of overlaps among pathway enrichment results using EnrichmentMap (EM) in Cytoscape.

Draws a network representation of overlaps among pathway enrichment results using EnrichmentMap (EM) in Cytoscape.

plotFemap(
    gmtFile,
    resultsFolder,
    pvalue = 1,
    qvalue = 1,
    formSim = "COMBINED",
    edgeSim = 0.375,
    combSim = 0.5,
    chartData = "NES_VALUE",
    clustAlg = "MCL",
    clustWords = 3,
    hideNodeLabels = FALSE,
    netName = "generic",
    netFile = "png"
)

Arguments

gmtFile
(char) absolute path to GMT file (generated via writePathways)

resultsFolder
(char) absolute path to folder with fedup results (generated via writeFemap)

pvalue
(numeric) pvalue cutoff (value between 0 and 1; default 1)

qvalue
(numeric) qvalue cutoff (value between 0 and 1; default 1)

formSim
(character) formula to calculate similarity score (one of OVERLAP, JACCARD, COMBINED; default COMBINED)

edgeSim
(numeric) edge similarity score cutoff (value between 0 and 1; default 0.375)

combSim
(numeric) when coefficients=COMBINED this parameter is used to determine what percentage to use for JACCARD and OVERLAP when combining their value (value between 0 to 1; default 0.5)

chartData
(char) node chart data (one of NES_VALUE, P_VALUE, FDR_VALUE, PHENOTYPES, DATA_SET, EXPRESSION_SET, or NONE; default NES_VALUE)

clustAlg
(character) clusterMaker algorith (one of AFFINITY_PROPAGATION, CLUSTER_FIZZIFIER, GLAY, CONNECTED_COMPONENTS, MCL, SCPS; default MCL)

clustWords
(integer) maximum words to include in autoAnnotate cluster label (default 3)

hideNodeLabels
(logical) if TRUE hides the node label in the EM; cluster labels generated via AutoAnnotate remain visible

netName
(char) name for EM in Cytoscape (default generic)

netFile
(char) name of output image (supports png, pdf, svg, jpeg image formats)

Value

File name of image to which the network is exported and an open session of Cytoscape (side effect of plotting EM). NULL if Cytoscape is not running locally.

Examples

# Load example data
data(geneDouble)
data(pathwaysGMT)
# Run fedup
fedupRes <- runFedup(geneDouble, pathwaysGMT)

#> Running fedup with:
#>  => 2 test set(s)
#>   + FASN_negative: 379 genes
#>   + FASN_positive: 298 genes
#>  => 17804 background genes
#>  => 1437 pathway annotations

#> All done!

# Write out results to temp folder
resultsFolder <- tempdir()
writeFemap(fedupRes, resultsFolder)

#> Wrote out EM-formatted fedup results file to /var/folders/mh/_0z2r5zj3k75yhtgm6l7xy3m0000gn/T//RtmpaRu10f/femap_FASN_negative.txt

#> Wrote out EM-formatted fedup results file to /var/folders/mh/_0z2r5zj3k75yhtgm6l7xy3m0000gn/T//RtmpaRu10f/femap_FASN_positive.txt

# Write out gmt formatted pathawy annotations to temp file
gmtFile <- tempfile("pathwaysGMT", fileext = ".gmt")
writePathways(pathwaysGMT, gmtFile)

#> Wrote out pathway gmt file to /var/folders/mh/_0z2r5zj3k75yhtgm6l7xy3m0000gn/T//RtmpaRu10f/pathwaysGMT84bf2509df1b.gmt

# Plot enrichment map
netFile <- tempfile("fedup_EM", fileext = ".png")
plotFemap(
    gmtFile = gmtFile,
    resultsFolder = resultsFolder,
    qvalue = 0.05,
    hideNodeLabels = TRUE,
    netName = "fedup_EM",
    netFile = netFile
)

#> Oh no! I can't find Cytoscape. RCy3 can not continue!
#> Please check that Cytoscape is running, CyREST is installed and your base.url parameter is correct.

#> NULL